Chore: moves transaction status to types
